Changed Float type to f32

+++ src/types.rs
@@ -1,5 +1,5 @@
 // (c) 2021 TVWS-Project
 // PAWS types
 
-pub type Float = f64;
+pub type Float = f32;
 pub type Int = i32;
